Add support for SAML RelayState
Realme supports accepting a relay state parameter (string < 80 chars) as
part of the SAML request. Realme will (in accordance with the SAML spec)
return the relay state unaltered with the response.

This change adds optional support for RelayState. Nothing changes about
existing functionality if you choose not to use RelayState.