Allow users to opt-in to SAML validation errors from ruby-saml (#19)
* Allow users to opt-in to SAML validation errors from ruby-saml

Getting exceptions from SAML validation errors can be really useful for
debugging, especially near the start of a project or when working with a
new IdP implementation. This opt-in setting allows users to choose to
see them.

The example in the docs recommends turning them on for Rails development
env only.

This change is backwards compatible. If the option is not specified it
defaults to off (which is the behaviour before this change)

* Update README.md

Co-authored-by: Ben <hapiben@users.noreply.github.com>

Co-authored-by: Ben <hapiben@users.noreply.github.com>