Add support for Realme Context Mapping Service (#18)
RCMS Details:
https://developers.realme.govt.nz/how-realme-works/whats-realme-rcms/

RCMS is an optional service that an app can enable on Realme if the app
needs to exchange user IDs (FLTs) with another app using Realme. RCMS
helps preserve the privacy of the user in this situation.

This change allows this gem to support RCMS if it is enabled. If RCMS is
not enabled then everything functions as it did before this change.

If RCMS is enabled then the RCMS Login Access Token (LAT) will be in
`request.env['omniauth.auth']['credentials']`.