Add support for SAML RelayState (#22)
* Implement standard OmniAuth error handling as an opt-in behaviour

When the appropriate option is set in the config, the strategy will now
use OmniAuth error reporting behaviour i.e. it will redirect to the
OmniAuth error Rack app.

This behaviour is currently opt-in to not break existing installations.

* Add support for SAML RelayState

Realme supports accepting a relay state parameter (string < 80 chars) as
part of the SAML request. Realme will (in accordance with the SAML spec)
return the relay state unaltered with the response.

This change adds optional support for RelayState. Nothing changes about
existing functionality if you choose not to use RelayState.